待機状態を設定

+++ prac_cgi/toggle.cgi
@@ -4,7 +4,13 @@ require 'fileutils'
 
 ASAKAI_FILE = '/tmp/asakai_skip.txt'
 
-if File.exist?(ASAKAI_FILE)
+env = ENV
+
+line = STDIN.gets
+
+if env["REQUEST_METHOD"] == "GET"
+  asakai = "朝会待機状態です"
+elsif env["REQUEST_METHOD"] == "POST" && line && line.match(/toggle=ON/) && File.exist?(ASAKAI_FILE)
   File.delete(ASAKAI_FILE)
   asakai = "朝会を行いました"
 else
@@ -22,6 +28,14 @@ puts <<END_CONTENT
 </head>
 <body>
 <h1 style='color: orange'>morning meeting</h1>
+<form action='/toggle_asakai.cgi' method='post'>
+<p>
+<input type="checkbox" name="toggle" value="ON" >朝会開始
+</p>
+<p>
+<input type='submit' value='送信'>
+</p>
+</form>
 <p>#{asakai}</p>
 </body>
 </html>
